IEP-552: Review comments added
