--- bundles/com.espressif.idf.debug.gdbjtag.openocd/src/com/espressif/idf/debug/gdbjtag/openocd/preferences/messages.properties
@@ -1,2 +1,0 @@
-GDBServerTimeoutPage_Description=Set GDB Server Launch Timeout
-GDBServerTimeoutPage_TimeoutField=Timeout for gdb server launch timeout
