+++ bundles/com.espressif.idf.serial.monitor/META-INF/MANIFEST.MF
@@ -17,7 +17,6 @@ Bundle-ActivationPolicy: lazy
 Export-Package: com.espressif.idf.serial.monitor,
  com.espressif.idf.serial.monitor.core,
  com.espressif.idf.serial.monitor.handlers,
- com.espressif.idf.serial.monitor.preferences,
  com.espressif.idf.serial.monitor.server
 Bundle-ClassPath: lib/javax.websocket-client-api-1.1.jar,
  lib/tyrus-standalone-client-1.9.jar,
