+++ bundles/com.espressif.idf.ui/src/com/espressif/idf/ui/preferences/messages.properties
@@ -1,1 +1,4 @@
 EspresssifPreferencesPage_IDFSpecificPrefs=ESP-IDF Specific Preferences.
+GDBServerTimeoutPage_TimeoutField=Timeout for gdb server launch timeout
+SerialMonitorPage_Field_NumberOfCharsInLine=Console Line Width (maximum characters):
+SerialMonitorPage_Field_NumberOfLines=Limit Console Output (number of lines):
